move serve tests
