+++ .buildkite/core.rayci.yml
@@ -389,9 +389,9 @@ steps:
     instance_type: medium
     commands:
       - bazel run //ci/ray_ci:build_in_docker -- docker --platform cpu
-        --canonical-tag test_container
-      - docker build --progress=plain --build-arg BASE_IMAGE="rayproject/ray:test_container"
-        -t rayproject/ray:runtime_env_container -f ci/docker/runtime_env_container/Dockerfile .
+        --canonical-tag test_image_uri
+      - docker build --progress=plain --build-arg BASE_IMAGE="rayproject/ray:test_image_uri"
+        -t rayproject/ray:runtime_env_image_uri -f ci/docker/runtime_env_image_uri/Dockerfile .
       - bazel run //ci/ray_ci:test_in_docker -- //python/ray/tests/... core --only-tags container --install-mask llm,rllib,serve,train,tune,workflow
     depends_on:
       - manylinux
