+++ python/ray/serve/tests/runtime_env_image_uri/test_serve_basic.py
