+++ python/ray/tests/runtime_env_image_uri/test_with_env_vars.py
