+++ python/ray/tests/test_runtime_env_image_uri.py
@@ -9,7 +9,7 @@ from ray.tests.conftest_docker import run_in_container, NESTED_IMAGE_NAME
 
 
 # NOTE(zcin): The actual test code are in python scripts under
-# python/ray/tests/runtime_env_container. The scripts are copied over to
+# python/ray/tests/runtime_env_image_uri. The scripts are copied over to
 # the docker container that's started by the `podman_docker_cluster`
 # fixture, so that the tests can be run by invoking the test scripts
 # using `python test.py` from within the pytests in this file
@@ -99,34 +99,10 @@ def test_worker_exit_intended_system_exit_and_user_error(
 
 
 @pytest.mark.skipif(sys.platform != "linux", reason="Only works on Linux.")
-@pytest.mark.parametrize("use_image_uri_api", [True, False])
-def test_serve_basic(podman_docker_cluster, use_image_uri_api):
-    """Test Serve deployment."""
-
-    container_id = podman_docker_cluster
-    cmd = ["python", "tests/test_serve_basic.py", "--image", NESTED_IMAGE_NAME]
-    if use_image_uri_api:
-        cmd.append("--use-image-uri-api")
-    run_in_container([cmd], container_id)
-
-
-@pytest.mark.skipif(sys.platform != "linux", reason="Only works on Linux.")
 def test_job(podman_docker_cluster):
 
     container_id = podman_docker_cluster
     cmd = ["python", "tests/test_job.py", "--image", NESTED_IMAGE_NAME]
-    run_in_container([cmd], container_id)
-
-
-@pytest.mark.skipif(sys.platform != "linux", reason="Only works on Linux.")
-@pytest.mark.parametrize("use_image_uri_api", [True, False])
-def test_serve_telemetry(podman_docker_cluster, use_image_uri_api):
-    """Test Serve deployment telemetry."""
-
-    container_id = podman_docker_cluster
-    cmd = ["python", "tests/test_serve_telemetry.py", "--image", NESTED_IMAGE_NAME]
-    if use_image_uri_api:
-        cmd.append("--use-image-uri-api")
     run_in_container([cmd], container_id)
